lustre: rename ops to owner
Now that portals_handle_ops contains only a char*,
it is functioning primarily to identify the owner of each handle.
So change the name to h_owner, and the type to const char*.

Note: this h_owner is now quite different from the similar h_owner
in the server code.  When server code is merged the
"med" pointer will be stored in the "mfd" and validated separately.

WC-bug-id: https://jira.whamcloud.com/browse/LU-12542
Lustre-commit: 1a9aafbf6317 ("LU-12542 handle: rename ops to owner")
Signed-off-by: NeilBrown <neilb@suse.com>
Reviewed-on: https://review.whamcloud.com/35798
Reviewed-by: Shaun Tancheff <stancheff@cray.com>
Reviewed-by: Neil Brown <neilb@suse.de>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>