lustre: replace lprocfs_climp_check()
The usage pattern for lprocfs_climp_check() is clumsy.
It must be paired with up_read of obd->u.cli.cl_sem, but
not doing this does not produce a compile-time error.
The 'import' should not be dereferenced before the CHECK, or
used after the EXIT, but sometimes it is.

Replace it with a structure macro/statement:

with_obd_imp_lock(obd, imp, rc) {
	statements;
}

statements are protected by the semaphore and only run if imp can be
set to a non-NULL pointer.
rc can be changed by the statements, and should be returned
afterwards as it may have been set to -ENODEV.

Errors fixed with this patch:
- some code tested u.cli.cl_import no-NULL even after
  lprocfs_climp_check()
- some code dereferences cl_import before calling
  lprocfs_climp_check()
- short_io_bytes_store() and max_procs_in_flight_store() don't access
  the import, so don't need lprocfs_climp_check()
- lprocfs_rd_import() set count to an error before 'goto out'
  which would free memory of length "count+1".
- lprocfs_rd_import() also called ptlrpc_recover_import()
  on the imp *after* dropping the semaphore.

WC-bug-id: https://jira.whamcloud.com/browse/LU-9855
Lustre-commit: f4ce6b568b52 ("LU-9855 lustre: replace LPROCFS_CLIMP_CHECK()")
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Reviewed-on: https://review.whamcloud.com/36956
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>
Reviewed-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>